Code/requirements.txt
argparse
easydict
h5py
matplotlib
numpy
open3d==0.9
opencv-python
pyyaml
scipy
tensorboardX
timm==0.4.5 
tqdm
transforms3d
termcolor
scikit-learn